Split DaemonRpc up into multiple parts
